Ispravke kod pisanja koda
