+++ src/components/CreateTodo.vue
@@ -1,65 +1,66 @@
 <template>
-    <div v-if="loggedIn!=='null'" align="center">
-        <input
-            autofocus autocomplete="off"
-            placeholder="What needs to be done?"
-            v-model="newText">
-        <div>
-        Priority:
-        <select style="margin: 10px;" v-model="priority">
-            <option v-for="option in options" v-bind:key="option">
-              {{ option }}
-            </option>
-        </select>
-        <div>
-          <button style="width:30%" @click="store()">Add task</button>
-        </div>
-      <p v-if="errors">
-          <b style="color:red">Fields cant be empty</b>
-        </p>
-      </div>
-        
-    </div>
+	<div v-if="loggedIn!=='null'" align="center">
+		<input
+			autofocus autocomplete="off"
+			placeholder="What needs to be done?"
+			v-model="newText">
+		<div>
+			Priority:
+			<select style="margin: 10px;" v-model="priority">
+				<option v-for="option in options" v-bind:key="option">
+					{{ option }}
+				</option>
+			</select>
+			<div>
+				<button style="width:30%" @click="store()">Add task</button>
+			</div>
+			<p>
+				<b style="color:red" id="errors"></b>
+			</p>
+	  </div>
+		
+	</div>
 </template>
+
 <script>
-    function Crud({ priority, text, id, completed}) {
-        this.priority = priority;
-        this.text = text;
-        this.id = id;
-        this.completed = completed;
-    }
-    import axios from 'axios';
-    import service from '../js/TaskService.js';
-    export default {
-        data() {
-            return {
-            newText: null,
-            input: {},
-            priority:'',
-            errors:false,
-            loggedIn:localStorage.getItem('token'),
-            options: [
-                'low',
-                'medium',
-                'high'
-                ]
-            }
-        },
-        methods:{
-            store() {
-                if(this.newText==null || this.newText=='' || this.priority==''){
-                    this.errors=true;
-                    return;
-                }else{
-                    this.errors=false;
-                }
-                var data = {};
-                service.addTask(this.newText, this.priority).then(d => {
-                    console.log(d);
-                    this.$emit('add',new Crud(d['data']["task"]))
-                });
-            },
-            
-        }
-    }
+function Crud({ priority, text, id, completed}) {
+	this.priority = priority;
+	this.text = text;
+	this.id = id;
+	this.completed = completed;
+}
+import axios from 'axios';
+import { taskService } from '../js/TaskService.js';
+export default {
+  data () {
+	  return {
+		  newText: null,
+		  input: {},
+		  priority:'',
+		  errors:false,
+		  loggedIn:localStorage.getItem('token'),
+		  options: [
+			  'low',
+			  'medium',
+			  'high'
+		  ]
+	  }
+  },
+
+  methods: {
+		
+	  store () {
+		  if (!this.newText || !this.priority) {
+				document.getElementById("errors").innerText = "Fields cant be empty"
+			  return;
+			} else { 
+				document.getElementById("errors").innerText = ""
+			}
+			taskService.addTask(this.newText, this.priority)
+				.then(d => {
+			  	this.$emit('add',new Crud(d['data']["task"]))
+		  	});
+	  }
+  }
+}
 </script>
