+++ src/components/Login.vue
@@ -1,0 +1,40 @@
+<template>
+	<div id="login" align="center">
+		<h1>Login</h1>
+		<input type="text" name="username" v-model="input.username" placeholder="Username" />
+		<input type="password" name="password" v-model="input.password" placeholder="Password" />
+		<button type="button" v-on:click="login()">Login</button>
+		<button align="left" type="button" v-on:click="showRegister()">Register</button>
+		<p>
+			<b style="color:red" id="errors"></b>
+    	</p>
+	</div>
+</template>
+
+<script>
+import { loginService } from '../js/LoginService.js';
+export default {
+	data () {
+		return {
+			input: {},
+		}
+	},
+	methods:{
+
+		login () {
+			loginService.login(this.input.username, this.input.password)
+			.then((data) => {
+				if (!data) {
+					document.getElementById("errors").innerText = "Wrong username/password!"
+				} else {
+					this.$router.replace('/');
+				}
+			});
+		},
+
+		showRegister () {
+			this.$router.push('/register');
+		}
+	}
+}
+</script>
